+++ src/main/java/dev/emortal/minestom/blocksumo/game/PlayerTags.java
@@ -8,8 +8,8 @@ public final class PlayerTags {
     public static final @NotNull Tag<Long> LAST_DAMAGE_TIME = Tag.Long("last_damage_timestamp");
     public static final @NotNull Tag<TeamColor> TEAM_COLOR = Tag.String("team_color").map(TeamColor::valueOf, TeamColor::name);
     public static final @NotNull Tag<Byte> LIVES = Tag.Byte("lives");
-    public static final @NotNull Tag<Byte> KILLS = Tag.Byte("kills");
-    public static final @NotNull Tag<Byte> FINAL_KILLS = Tag.Byte("final_kills");
+    public static final @NotNull Tag<Integer> KILLS = Tag.Integer("kills");
+    public static final @NotNull Tag<Integer> FINAL_KILLS = Tag.Integer("final_kills");
     public static final @NotNull Tag<Boolean> CAN_BE_HIT = Tag.Boolean("can_be_hit");
     public static final @NotNull Tag<Long> SPAWN_PROTECTION_TIME = Tag.Long("spawn_protection_time");
